divided player.rs into multiple files
