+++ src/main_menu/mod.rs
@@ -2,8 +2,8 @@ use bevy::{app::AppExit, prelude::*};
 
 use crate::{
     audio::{GameAudioOptions, GameAudioState},
+    game_states::GameStates,
     pause_menu::button::{UIButton, HOVERED_BUTTON, NORMAL_BUTTON, PRESSED_BUTTON},
-    states::GameStates,
 };
 mod ui;
