+++ src/player/lives_counter.rs
