+++ src/toast/mod.rs
@@ -8,7 +8,7 @@ use bevy_tweening::{
 };
 
 use crate::cheat_codes::CheatCodeResource;
-use crate::runner::CollectedChars;
+use crate::player::CollectedChars;
 
 pub struct ShowToast {
     pub value: String,
